requirements.txt
numpy
pillow
torch
fastapi
uvicorn
pydantic
